Make venv change work on Darwin
This patch builds on the previous to make prefix resolution work
correctly inside virtual environments on Darwin. The --inherit-argv0
mechanism is insufficient, as on Darwin, Python uses the
_NSGetExecutablePath() call to work out the actual path to its binary.
Since this resolves to the real unwrapped binary, --inherit-argv0 is
little use.

This patch adds a new variant, --inherit-argv0-path, which resolves the
full path before passing it to argv[0]. On Darwin this is done using
_NSGetExecutablePath(), on Linux by reading /proc/self/exe. This
bypasses the real_executable lookup, as Python gives precedence to
argv[0] when it appears to be a path (contains a slash).

While not strictly necessary on Linux (as --inherit-argv0 works anyway),
the change is made to both for consistency.